+++ CHANGELOG.md
@@ -14,6 +14,7 @@ You can also check the
 - Fixes
   - Chart more button is now correctly hidden when there are no actions
     available
+  - Most recent value is now correctly resolved in interactive filters
 
 # 5.9.0 - 2025-07-28